ham: power: Disable input boost while screen is off
For unknown reasons the kernel boost driver interprets
certain events as input boost. This becomes appearent
because not all music players use wakelocks any more
(Google's Play Music being a prominent example).

While it is great that our cpu is now able to use deeper
sleep states during offloaded media playback most of the
time, it is not so great that the benefits are outweighed
by unnessecarily high freqs while ghost-boosted.  To work
around this, disable input boost while we don't need it
anyway.

Set proper permissions on the sysfs node as well.

Also remove "libqti-iop-client.so" check as we dont have this lib included

Change-Id: I6cc5b61c5879ca4bd603ccd4930a05b70854acc2
Signed-off-by: ShubhamB <shubhamprince111@gmail.com>
Signed-off-by: Srijith <srijithdamerla2001@gmail.com>